refactor: move add repository API back to ProjectController and remove checkURL in add repo API
